fixed some errors and changed dataset
We have updated the code as requested. About your last  question the batch size with the full length of the dataset and the abscence of a validation split were mistakes on our part we hope we fixed. Finally, we tested various datasets but we could not reproduce a better result than the double moon example shown in the paper: for this reason we have fallen back to the old dataset from sklearn. Is it okay? Maybe we should implement it from scratch without the help of the library? Please let us know.
Best regards,
Zannini Quirini Giorgio, Mandolini Giorgio Maria, Sanna Daniele