RealNVP (#188)
* RealNVP

We are three italian data scientists to be and we would love to know what you think if our project could be a good example!

* Update and rename Density estimation using RealNVP to Density estimation using RealNVP.py

Good evening! We have made the changes you requested, if there is any other mistake or if it is not well formatted please let us know.
About the dataset, we have looked for a more realistic one for the density estimation task; we found a dataset with multiple clusters that we think might more accurately represent a real-life scenario.
Best regards,
Sanna Daniele, Zannini Quirini Giorgio, Mandolini Giorgio Maria

* Copyediting

* Fixed code

Good evening,
We have modified the code according to your suggestions. Should there be anything we need to change, please let us know.
Best regards,
Mandolini Giorgio Maria, Sanna Daniele, Zannini Quirini Giorgio

* fetched old copyedit, changed self.call(x) to self(x)

Sorry we somehow missed that copyedit. Here is the corrected version.

* fixed some errors and changed dataset

We have updated the code as requested. About your last  question the batch size with the full length of the dataset and the abscence of a validation split were mistakes on our part we hope we fixed. Finally, we tested various datasets but we could not reproduce a better result than the double moon example shown in the paper: for this reason we have fallen back to the old dataset from sklearn. Is it okay? Maybe we should implement it from scratch without the help of the library? Please let us know.
Best regards,
Zannini Quirini Giorgio, Mandolini Giorgio Maria, Sanna Daniele

* Fixed small mistake on commit

Sorry that was an old version of the code, this should be the right one.

* fixed the loss tracker

Thanks for the suggestions which are helping us learning aspects of Keras we didn't know about! About the part of ```z, _ = model(normalized_data)```, we don't understand how to use ```predict``` here since it gives some weird result if we use it.

* added missing comment on `metrics`

Sorry about that!

* Update Density estimation using RealNVP.py

* self.call to self

We are very sorry we had some trouble with the versions. It won't happen again.

* Rename Density estimation using RealNVP.py to real_nvp.py

* removed summary

* deleted summary

* fetched old copyedit

* Copyedits

* uploaded real_nvp.md

* uploaded real_nvp.ipynb

* uploaded real_nvp images

Co-authored-by: François Chollet <francois.chollet@gmail.com>