Update serialization_and_saving.py (#318)
* Update serialization_and_saving.ipynb

Trivial bracket typo.
Changed the order
I can see the sentence 'Below is an example of ... from he SavedModel format without overwriting the config methods.' on the page, but somehow I can't find it in the sourcecode. Hope 'he -> the'.

* Update serialization_and_saving.md

Trivial bracket typo.
Changed the order
I can see the sentence 'Below is an example of ... from he SavedModel format without overwriting the config methods.' on the page, but somehow I can't find it in the sourcecode. Hope 'he -> the'.

* Update serialization_and_saving.py