Update oxford_pets_image_segmentation.py (#351)
* Update oxford_pets_image_segmentation.py

For the two curl commands in lines 13 and 14, using URLs with the http protocol downloads files that are not recognized by the subsequent tar commands as tar archives. Using the https protocol solves the issue.

* Update oxford_pets_image_segmentation.md

Fix URLs on lines 18 and 19 to use the https protocol. Using the regular http protocol downloads files that are not recognized by the subsequent tar commands as tar archives

* Update oxford_pets_image_segmentation.ipynb

Fix URLs on lines 34 and 35 to use the https protocol. Using the regular http protocol downloads files that are not recognized by the subsequent tar commands as tar archives

* Update oxford_pets_image_segmentation.ipynb

Remove extraneous newline at the end of file