Update Save and load Keras subclassed models
The guide states that to save/load custom layer or a subclassed model, the get_config and optionally from_config methods should be overwritten. However, in the case of subclassed model the definition of from_config is needed and not optional. Updated to reflect the need of definition of from_config for subclassed models.