Correct description of logs on epoch end
Of what I understand of the code, and it's coherent with the output 
in the current file, is that the loss value in the logs of the 
funciton `on_epoch_end` is that of the last step of the epoch rather 
than the mean of the losses of its steps.

I guess the intention what to return the average loss of the epoch's
steps. However, since it's not what was implemented the documentation
shouldn't be misleading.