Update ner_transformers.py (#846)
* Update ner_transformers.py

The bug was found in Issue #801. During the dataset export (lines 137-150), an extra line was added at the end of the dataset files (train and validation), where the number of tokens was 0, and the rest was an empty string. Fixed by adding an additional check to verify that the number of tokens is larger than 0 before writing to the files.

* Update ner_transformers.md

Bugfix in md file.

* Update ner_transformers.ipynb

Bugfix jupyter notebook