Fixes Caltech dataset broken link (#1165)
* Fixes broken link for Caltech Dataset

Fixes broken link for Caltech Dataset. Patch clone for #1066. 
Trying to copy the changes to .ipynb and .md on behalf of leaving team.

* Copied the changes to md and ipynb

Copied the changes to md and ipynb.  also updated the download link for caltech zip file  in code.
path_to_downloaded_file

* copied the changes to .ipynb and .md

copied the changes to .ipynb and .md

* Updated path_to_download_file in the .py file

Updated path_to_download_file in the .py file