Updated getting_started.py (#1248)
* Added responsiveness to landing page images.

Some images of the landing page are not responsive on some devices. So I played with the CSS of the landing page and found out that `min-height: 20rem` works better than `min-height: 25rem`. That's why I changed it to 20rem.

* Updated getting_started.py

I have done small improvements in the guide such as spelling mistakes (imdb was written as imdb somewhere). I have also added comments so that beginners can easily understand the code.

* Updated getting_started.py

I have made the changes as suggested by the reviewer.

* Updated getting_started.py

Made the changes as suggested by the reviewer.

* Created using Colaboratory

* Reverting changes

* generating ipynb and md files

---------

Co-authored-by: Neeshamraghav012 <neeshamraghav0@gmail.com>