Keras 3 ner_transformers (#1610)
Tensorflow only with preprocessing and some modeling ops in tf, but
it should not be too hard to port this to multi backend with tf-only
preprocessing in the future.

I believe there was a bad bug in this examples, where it computed a
loss of softmax outputs with from_logits=True. Fixed.