Port sentence_embeddings_with_sbert to Keras 3
Only tf backend for now, and `jit_compile=False` to avoid a
XLA error with variables on the wrong device.
https://www.tensorflow.org/xla/known_issues#tfvariable_on_a_different_device

The long term fix is probably to move all preprocessing out of the call
graph of these models (no `Input(dtype="string")`), but for now we can
do the simple fix.