Correct LoRA weights merging
Correction of the merging code between the model's original layer weights and the LoRA model weights.
This respect the principle of LoRA to dispose of the LoRA layers once we don't plan on training it more bur more importantly allows us to save and load the model as a ".keras" file.