Move Keras 2 apis to a constant url
We had the idea of a rolling Keras 2 api url, that would append to
the set of files we publish on keras.io. It's broken for a couple
reasons:

1. The site might change between tf version, and we don't want to
   serve old versions of pages from the site hoping that they match up.
2. We should have a stable url for the latest keras 2 api for things
   like SEO. We don't have that!

We should only serve what we can build from the current version of
the site. If we want rolling selectable API versions, we should do
so by extending our site build to be able to build all Keras 2 versions
we care about when we render.