Updates to "Customizing fit() with JAX" guide.
The starting point for these changes was to use `stateless_compute_loss` instead of `compute_loss` for consistency, since we state that everything in `train_step` should be stateless.
- This in turn required plumbing `metrics_variable` and `sample_weight`.
- This required adding `keras.utils.unpack_x_y_sample_weight`.

These changes make the guide slightly longer, but the code is more generic and closer to the Keras default implementation.

Also fixed issue where `test_step` was not returning the metrics correctly.